+++ app/[locale]/(auth)/sign-in/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import EmailSignInForm from "@app/(auth)/sign-in/_components/sign-in-form";
+import EmailSignInForm from "@app/[locale]/(auth)/sign-in/_components/sign-in-form";
 
 export default function SignInPage() {
   return (
